Apply validation to autofill address pattern
